+++ app/support/search/parser.ts
@@ -1,4 +1,6 @@
 import XRegExp from 'xregexp';
+import { flow } from 'lodash';
+import config from 'config';
 
 import { normalizeText } from './norm';
 import {
@@ -6,12 +8,14 @@ import {
   listConditions,
   ScopeStart,
   Pipe,
+  Plus,
   Condition,
   Text,
-  AnyText,
   InScope,
   trimText,
-  Token
+  Token,
+  SeqTexts,
+  AnyText
 } from './query-tokens';
 
 // -?(scope:)?(double-quoted-string|string)
@@ -19,6 +23,7 @@ const tokenRe = XRegExp(
   `
   (?:
     (?<pipe> \\|) |
+    (?<plus> \\+) |
     (?:
       (?<exclude> -)?
       (?:(?<cond> [\\w-]+):)?
@@ -32,7 +37,11 @@ const tokenRe = XRegExp(
   'gx'
 );
 
-export function parseQuery(query: string) {
+export type ParseQueryOptions = {
+  minPrefixLength: number
+}
+
+export function parseQuery(query: string, { minPrefixLength }: ParseQueryOptions = config.search) {
   // 1-st run: Split the query string into tokens
 
   const tokens = [] as Token[];
@@ -42,6 +51,11 @@ export function parseQuery(query: string) {
 
     if (match.pipe) {
       tokens.push(new Pipe());
+      return;
+    }
+
+    if (match.plus) {
+      tokens.push(new Plus());
       return;
     }
 
@@ -76,43 +90,41 @@ export function parseQuery(query: string) {
       // (-)in-body:cat,mouse
       for (const [re, scope] of scopeStarts) {
         if (re.test(match.cond)) {
-          const inner = [];
-
           if (match.qstring) {
             // in-body:"cat mouse" => "cat mouse"
-            const phrase = new Text(
-              !!match.exclude,
-              true,
-              JSON.parse(match.qstring)
+            tokens.push(new InScope(scope,
+              new AnyText([
+                new Text(
+                  !!match.exclude,
+                  true,
+                  JSON.parse(match.qstring)
+                )
+              ]))
             );
-            inner.push(new AnyText([phrase]));
           } else {
             const words = (match.word as string)
               .split(',')
-              .map(trimText)
+              .map((w) => trimText(w, { minPrefixLength }))
               .filter(Boolean);
 
             if (!match.exclude) {
               // in-body:cat,mouse => cat || mouse
               const texts = words.map((word) => new Text(false, false, word));
-              inner.push(new AnyText(texts));
+              tokens.push(new InScope(scope, new AnyText(texts)));
             } else {
               // -in-body:cat,mouse => !cat && !mouse
-              const parts = words.map(
-                (word) => new AnyText([new Text(true, false, word)])
-              );
-              inner.push(...parts);
+              const texts = words.map((word) => new Text(true, false, word));
+              tokens.push(...texts.map((t) => new InScope(scope, new AnyText([t]))));
             }
           }
 
-          tokens.push(new InScope(scope, inner));
           return;
         }
       }
 
       // Scope not found, treat as raw text
       tokens.push(
-        new AnyText([new Text(!!match.exclude, false, trimText(raw))])
+        new AnyText([new Text(!!match.exclude, false, trimText(raw, { minPrefixLength }))])
       );
       return;
     }
@@ -123,15 +135,29 @@ export function parseQuery(query: string) {
         new Text(
           !!match.exclude,
           !!match.qstring,
-          match.qstring ? JSON.parse(match.qstring) : trimText(match.word)
+          match.qstring ? JSON.parse(match.qstring) : trimText(match.word, { minPrefixLength })
         )
       ])
     );
   });
 
-  // 2-nd run: Merge all "AnyText (Pipe AnyText)+" combinations into one AnyText.
-  // Result should not contain any Pipe's.
+  return flow([
+    joinByPipes,
+    joinByPluses,
+  ])(tokens) as Token[];
+}
 
+export function queryComplexity(tokens: Token[]) {
+  return tokens.reduce((acc, token) => acc + token.getComplexity(), 0);
+}
+
+// Token list post-processors
+
+/**
+ *  1-st run: Merge all "AnyText (Pipe AnyText)+" combinations into one
+ *  AnyText.Result should not contain any Pipe's.
+ */
+function joinByPipes(tokens: Token[]) {
   const result = [] as Token[];
   let prevToken = null;
 
@@ -144,7 +170,7 @@ export function parseQuery(query: string) {
 
       if (i < tokens.length - 1 && tokens[i + 1] instanceof AnyText) {
         // Next token is AnyText, join it with the prevToken
-        prevToken.texts.push(...(tokens[i + 1] as AnyText).texts);
+        prevToken.children.push(...(tokens[i + 1] as AnyText).children);
         // Jump over the joined token
         i++;
         continue;
@@ -158,6 +184,34 @@ export function parseQuery(query: string) {
   return result;
 }
 
-export function queryComplexity(tokens: Token[]) {
-  return tokens.reduce((acc, token) => acc + token.getComplexity(), 0);
+function joinByPluses(tokens: Token[]) {
+  const result = [] as Token[];
+  let prevToken = null;
+
+  for (let i = 0; i < tokens.length; i++) {
+    if (tokens[i] instanceof Plus) {
+      if (!prevToken || !(prevToken instanceof SeqTexts)) {
+        // Last inserted token is not an SeqTexts
+        continue;
+      }
+
+      if (i < tokens.length - 1 && tokens[i + 1] instanceof AnyText) {
+        // Next token is AnyText, join it with the prevToken
+        prevToken.children.push(tokens[i + 1] as AnyText);
+        // Jump over the joined token
+        i++;
+        continue;
+      }
+    } else {
+      prevToken = tokens[i];
+
+      if (prevToken instanceof AnyText) {
+        prevToken = new SeqTexts([prevToken]);
+      }
+
+      result.push(prevToken);
+    }
+  }
+
+  return result;
 }