+++ app/support/search/to-tsvector.ts
@@ -13,21 +13,31 @@ const ftsCfg = config.postgres.textSearchConfigName;
 export function toTSVector(text: string) {
   const vectors = tokenize(normalizeText(text || '')).map((token) => {
     if (token instanceof HashTag || token instanceof Mention) {
-      // Mentions and hashtags should be found by exact @-query or by regular word query
-      const exactText =
-        token instanceof HashTag ? token.text.replace(/[_-]/g, '') : token.text;
+      // Mentions and hashtags should be found by exact @/#-query or by regular word query
+      const exactText = token instanceof HashTag
+        ? token.text.replace(/[_-]/g, '') // join parts of hashtag to ignore separators
+        : token.text;
       return pgFormat(
-        `(to_tsvector(%L, %L)::text || ' ' || %L)::tsvector`,
+        `(to_tsvector_with_exact(%L, %L)::text || ' ' || %L)::tsvector`,
         ftsCfg,
-        token.text.replace(/[_-]+/g, ' '),
+        token.text.substring(1).replace(/[_-]+/g, ' '), // convert separated text to phrase
         `'${exactText}':1`
       );
-    } else if (token instanceof Link) {
-      return pgFormat(`to_tsvector(%L, %L)`, ftsCfg, linkToText(token));
     }
 
-    return pgFormat('to_tsvector(%L, %L)', ftsCfg, token.text);
-  });
+    if (token instanceof Link) {
+      return pgFormat(`to_tsvector_with_exact(%L, %L)`, ftsCfg, linkToText(token));
+    }
 
-  return vectors.length > 0 ? `(${vectors.join('||')})` : `to_tsvector('')`;
+    const trimmedText = token.text.trim();
+    return trimmedText && pgFormat('to_tsvector_with_exact(%L, %L)', ftsCfg, trimmedText);
+  }).filter(Boolean);
+
+  if (vectors.length === 0) {
+    return `''::tsvector`;
+  } else if (vectors.length === 1) {
+    return vectors[0];
+  }
+
+  return `(${vectors.join(' || ')})`;
 }